+++ Frends.Pgp.DecryptFile/Frends.Pgp.DecryptFile/Definitions/Input.cs
@@ -30,11 +30,38 @@ public class Input
     public int DecryptBufferSize { get; init; } = 64;
 
     /// <summary>
+    /// Source of the private key - file path or string.
+    /// </summary>
+    /// <example>PrivateKeySource.File</example>
+    public PrivateKeySource PrivateKeySource { get; set; } = PrivateKeySource.File;
+
+    /// <summary>
     /// Path to the private key file.
     /// </summary>
     /// <example>C:\keys\my_key.asc</example>
+    [UIHint(nameof(PrivateKeySource), "", PrivateKeySource.File)]
     [DisplayFormat(DataFormatString = "Text")]
     public string PrivateKeyPath { get; set; }
+
+    /// <summary>
+    /// Private key as a string, including the BEGIN/END PGP PRIVATE KEY BLOCK lines.
+    /// </summary>
+    /// <example>
+    /// -----BEGIN PGP PRIVATE KEY BLOCK-----
+    /// ...
+    /// -----END PGP PRIVATE KEY BLOCK-----
+    /// </example>
+    [UIHint(nameof(PrivateKeySource), "", PrivateKeySource.String)]
+    [DisplayFormat(DataFormatString = "Text")]
+    [PasswordPropertyText]
+    public string PrivateKeyString { get; set; }
+
+    /// <summary>
+    /// Encoding used for the private key passphrase. Defaults to Utf8.
+    /// Use Legacy if the passphrase contains non-ASCII characters and the key was created with an older PGP tool.
+    /// </summary>
+    /// <example>PassphraseEncoding.Utf8</example>
+    public PassphraseEncoding PassphraseEncoding { get; set; } = PassphraseEncoding.Utf8;
 
     /// <summary>
     /// Passphrase for the private key.